Implemented the bit translation table for setDigitRaw()/setRow(). Added setRow() as an alias for setDigitRaw(). Minor fixes.
The setDigitRaw() method now applies the bit translation described in the
documentation. A value of B00000001 now lights the LED connected to SEG A,
B10000000 lights SEG DP. Segments A-G + DP now correspond to the bit indices
0-7. This should make automatic adressing of the segments easier.
The constructor now also sets the brightness to full because the MAX7219 starts
out at lowest brightness when enabled after a reset.